[ADM-877][frontend]: feat: restructure files and simply include `+` button.
