+++ frontend/src/constants/resources.ts
@@ -423,3 +423,5 @@ export const AUTHORIZE_ORGANIZATION_LINK =
 
 export const DEFAULT_MONTH_INTERVAL_DAYS = 30;
 export const TIME_RANGE_TITLE = 'Time range settings';
+
+export const ADD_TIME_RANGE_BUTTON_TEXT = 'New time range';
