+++ frontend/src/containers/ConfigStep/DateRangePicker/types.ts
@@ -1,0 +1,5 @@
+export interface IRangePickerProps {
+  startDate: string | null;
+  endDate: string | null;
+  index: number;
+}
