+++ frontend/src/containers/MetricsStepper/index.tsx
@@ -127,7 +127,8 @@ const MetricsStepper = () => {
         { isShow: isShowSourceControl, isValid: isSourceControlVerified },
       ];
       const activeNextButtonValidityOptions = nextButtonValidityOptions.filter(({ isShow }) => isShow);
-      projectName && dateRange.startDate && dateRange.endDate && metrics.length
+      // todo refactor this to adapt new getInfo api design
+      projectName && dateRange && dateRange.length && metrics.length
         ? setIsDisableNextButton(!activeNextButtonValidityOptions.every(({ isValid }) => isValid))
         : setIsDisableNextButton(true);
     }
